+++ maas-client/client/src/main/java/com/netcracker/cloud/maas/client/api/MaaSException.java
@@ -1,7 +1,13 @@
 package com.netcracker.cloud.maas.client.api;
 
 public class MaaSException extends RuntimeException {
+
     public MaaSException(String format, Object...args) {
         super(String.format(format, args));
+    }
+
+    /** For a message that is already built and must not go through String.format. */
+    protected MaaSException(String message, Throwable cause) {
+        super(message, cause);
     }
 }